fix custom project
