+++ src/components/GlobalStyles/GlobalStyles.js
@@ -1,0 +1,12 @@
+import PropTypes from 'prop-types';
+import './GlobalStyles.scss';
+
+function GlobalStyles({ children }) {
+    return children;
+}
+
+GlobalStyles.propTypes = {
+    children: PropTypes.node.isRequired,
+};
+
+export default GlobalStyles;
