+++ src/index.js
@@ -5,12 +5,15 @@ import reportWebVitals from './reportWebVitals';
 import { Provider } from 'react-redux';
 import { store } from './app/store';
 import App from '~/App';
+import GlobalStyles from './components/GlobalStyles/GlobalStyles';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
     <React.StrictMode>
         <Provider store={store}>
-            <App />
+            <GlobalStyles>
+                <App />
+            </GlobalStyles>
         </Provider>
     </React.StrictMode>,
 );
