+++ src/pages/Publish/Signup/Signup.module.scss
@@ -1,6 +1,6 @@
-body {
-    background-image: radial-gradient(closest-side at 50% 135%, #ffffff 50%, #eceff1 100%);
-}
+// body {
+//     background-image: radial-gradient(closest-side at 50% 135%, #ffffff 50%, #eceff1 100%);
+// }
 .wrapper {
     width: 400px;
     height: 95vh;
